MUC Cleaner - Removed output namespace.
