+++ cleaner/muc/classes/index.php
@@ -22,7 +22,7 @@
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace cleaner_muc\output;
+namespace cleaner_muc;
 
 defined('MOODLE_INTERNAL') || die();
