+++ cleaner/muc/tests/phpunit/downloader_test.php
@@ -22,7 +22,7 @@
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-use cleaner_muc\output\downloader;
+use cleaner_muc\downloader;
 
 defined('MOODLE_INTERNAL') || die();
